# backend/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /orcamentos/{orcamentoId}/mensagens/{mensagemId}/visualizar:
    post:
      operationId: MensagensNegociacaoController_marcarComoVisualizada
      parameters:
        - name: orcamentoId
          required: true
          in: path
          schema:
            type: string
        - name: mensagemId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags: &ref_26
        - MensagensNegociacao
...
  /orcamentos/{orcamentoId}/mensagens/publico:
    get:
      operationId: MensagensNegociacaoController_findAllPublico
      parameters:
        - name: orcamentoId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_26
    post:
      operationId: MensagensNegociacaoController_createPublico
      deprecated: true
      summary: Descontinuado — use MensagemChat em /orcamentos-v2
      description: Fase 1 — escrita órfã; responde 410 Gone.
      parameters:
        - name: orcamentoId
          required: true
          in: path
          schema:
            type: string
      responses:
        '410':
          description: Gone — chat legado descontinuado
      tags: *ref_26
  /orcamentos/{orcamentoId}/mensagens/nao-visualizadas:
    get:
      operationId: MensagensNegociacaoController_findNaoVisualizadas
      parameters:
        - name: orcamentoId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_26
  /orcamentos/{orcamentoId}/mensagens/nao-visualizadas/count:
    get:
      operationId: MensagensNegociacaoController_countNaoVisualizadas
      parameters:
        - name: orcamentoId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_26
  /orcamentos/{orcamentoId}/mensagens:
    get:
      operationId: MensagensNegociacaoController_findAll
      parameters:
        - name: orcamentoId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_26
    post:
      operationId: MensagensNegociacaoController_create
      deprecated: true
      summary: Descontinuado — use MensagemChat em /orcamentos-v2
      description: >
        Fase 1: superfície órfã. Escritas respondem 410 Gone.
        Contrato canônico: MensagemChat via orcamentos-v2.
      parameters:
        - name: orcamentoId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMensagemNegociacaoDto'
      responses:
        '410':
          description: Gone — chat legado descontinuado; use MensagemChat
      tags: *ref_26
  /orcamentos/{orcamentoId}/mensagens/publico/{mensagemId}/visualizar:
    post:
      operationId: MensagensNegociacaoController_marcarComoVisualizadaPublico
      parameters:
        - name: orcamentoId
          required: true
          in: path
          schema:
            type: string
        - name: mensagemId
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags: *ref_26
  /orcamentos/{orcamentoId}/mensagens/{mensagemId}/anexo:
    post:
      operationId: MensagensNegociacaoController_uploadAnexo
      deprecated: true
      summary: Descontinuado — use MensagemChat em /orcamentos-v2
      description: Fase 1 — escrita órfã; responde 410 Gone.
      parameters:
        - name: orcamentoId
          required: true
          in: path
          schema:
            type: string
        - name: mensagemId
          required: true
          in: path
          schema:
            type: string
      responses:
        '410':
          description: Gone — chat legado descontinuado
      tags: *ref_26
...
  /orcamentos-v2/chat/{id}/mensagens/{mensagemId}/visualizar:
    post:
      operationId: OrcamentosV2Controller_marcarMensagemComoLida
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: mensagemId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Mensagem marcada como lida
        '404':
          description: Mensagem não encontrada
      security: *ref_30
      summary: Marcar mensagem como lida (autenticado)
      tags: *ref_31
...
  /orcamentos-v2/{id}/mensagens/publico:
    get:
      operationId: OrcamentosV2Controller_buscarMensagensChatPublico
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Mensagens encontradas
        '404':
          description: Orçamento não encontrado
      security: *ref_30
      summary: Buscar mensagens do chat público
      tags: *ref_31
    post:
      operationId: OrcamentosV2Controller_enviarMensagemChatPublico
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Mensagem enviada com sucesso
        '404':
          description: Orçamento não encontrado
      security: *ref_30
      summary: Enviar mensagem no chat público
      tags: *ref_31
  /orcamentos-v2/{id}/publico/mensagens/{mensagemId}/visualizar:
    post:
      operationId: OrcamentosV2Controller_marcarMensagemComoLidaPublico
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: mensagemId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Mensagem marcada como lida
        '404':
          description: Mensagem não encontrada
      security: *ref_30
      summary: Marcar mensagem como lida (chat público)
      tags: *ref_31
...
info:
  title: Comunikapp API
  description: |2-

            API do sistema Comunikapp - Comunicação Visual
            
            ## Módulos Disponíveis
            
            ### Orçamentos V2
            - Criação e gestão de orçamentos
            - Cálculo automático de custos
            - Integração com insumos e serviços
            
            ### Ordens de Serviço (OS)
            - Criação de OS derivadas de orçamentos
            - OS diretas/internas
            - Workflows de aprovação
            - Validação de estoque
            
            ### PCP (Planejamento e Controle de Produção)
            - Workflows configuráveis
            - Apontamentos de produção
            - Controle de qualidade
            
            ### Estoque
            - Gestão de insumos
            - Movimentações de estoque
            - Alertas de reposição
            
            ### Insumos
            - Catálogo de materiais
            - Preços e fornecedores
            - Categorização
            
            ## Autenticação
            
            Todas as APIs requerem autenticação JWT. Inclua o token no header:
            ```
            Authorization: Bearer <seu-token>
            ```
            
            ## Multi-Tenant
            
            O sistema é multi-tenant por loja. O lojaId é automaticamente 
            extraído do token JWT e aplicado a todas as operações.
            
            ## Versionamento
            
...
            
...
servers:
...
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Token JWT para autenticação
      in: header
  schemas:
...
    CreateMensagemNegociacaoDto:
      type: object
      properties: {}
...